python-serv/requirements.txt
black @ git+https://github.com/psf/black@9b161072c13c0ec32c9ca9bd48fad17f781a56d4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
PyJWT==1.4.2
requests==2.27.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
